The module is now able to calculate up to fib(184)
This commit implements a naive method to let fibdrv able to
calculate up to fib(184).
The idea is to split the result into 2 unsigned long long
number called upper and lower while calculating. Then, turn
the answer into a string and send it to user space through
copy_to_user.